requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
asn1crypto==1.5.1
asyncio==3.4.3
attrs==23.2.0
bech32==1.2.0
bip-utils==2.9.3
bitarray==2.9.2
Brownie==0.5.1
cbor2==5.6.4
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
coincurve==20.0.0
crcmod==1.7
Cython==0.29.37
cytoolz==0.12.3
dateparser==1.2.0
dydx-v4-client==0.1.0
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.9.0
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.4.0
eth-rlp==1.0.1
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.1.0
frozenlist==1.4.1
grpcio==1.65.1
grpcio-tools==1.62.2
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
mpmath==1.3.0
multidict==6.0.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parsimonious==0.10.0
patsy==0.5.6
protobuf==4.25.3
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pyunormalize==15.1.0
pywin32==306
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rlp==4.0.1
rpds-py==0.19.0
scipy==1.14.0
setuptools==71.0.1
six==1.16.0
sniffio==1.3.1
statsmodels==0.14.2
sympy==1.13.0
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
v4-proto==4.1.4
web3==6.20.0
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
